Move `auth/handshake.go` to `ws/handshake.go`
